* Fixed: Possible runtime assert
git-svn-id: https://svn.code.sf.net/p/codeblocks/code/trunk@9013 2a5c6006-c6dd-42ca-98ab-0921f2732cef

+++ src/plugins/codecompletion/insertclassmethoddlg.cpp
@@ -109,7 +109,7 @@ wxArrayString InsertClassMethodDlg::GetCode() const
     wxArrayString array;
     const wxCheckListBox* clb = XRCCTRL(*this, "chklstMethods", wxCheckListBox);
 
-    for (int i = 0; i < (int)clb->GetCount(); ++i)
+    for (unsigned int i = 0; i < clb->GetCount(); ++i)
     {
         if (clb->IsChecked(i))
         {
